refactor: create presentation service to make ICL call and format data for Mat Table
